+++ src/components/pacbio/PacbioRunWellEdit.vue
@@ -69,7 +69,7 @@
 // A lot of it could be moved to the store
 import { mapState, mapActions } from 'pinia'
 import PacbioRunWellComponents from '@/config/PacbioRunWellComponents'
-import { usePacbioRunCreateStore } from '@/stores/pacbioRunCreate'
+import { usePacbioRunCreateStore } from '@/stores/pacbioRunCreateV1.js'
 
 export default {
   name: 'WellModal',
