+++ tests/unit/components/pacbio/PacbioRunInfoEdit.spec.js
@@ -1,8 +1,8 @@
-import PacbioRunInfoEdit from '@/components/pacbio/PacbioRunInfoEdit'
-import { mount, createTestingPinia } from '@support/testHelper'
+import PacbioRunInfoEdit from '@/components/pacbio/PacbioRunInfoEdit.vue'
+import { mount, createTestingPinia } from '@support/testHelper.js'
 import { beforeEach, describe, expect } from 'vitest'
-import { PacbioInstrumentTypes } from '@/lib/PacbioInstrumentTypes'
-import { usePacbioRunCreateStore } from '@/stores/pacbioRunCreate'
+import { PacbioInstrumentTypes } from '@/lib/PacbioInstrumentTypes.js'
+import { usePacbioRunCreateStore } from '@/stores/pacbioRunCreateV1.js'
 
 // required as suggestion to remove the deprecated function
 // https://vue-test-utils.vuejs.org/api/options.html#attachtodocument
