+++ tests/unit/components/pacbio/PacbioRunWellDefaultEdit.spec.js
@@ -1,8 +1,8 @@
-import PacbioRunWellDefaultEdit from '@/components/pacbio/PacbioRunWellDefaultEdit'
-import { mount, createTestingPinia } from '@support/testHelper'
+import PacbioRunWellDefaultEdit from '@/components/pacbio/PacbioRunWellDefaultEdit.vue'
+import { mount, createTestingPinia } from '@support/testHelper.js'
 import { describe, expect, it } from 'vitest'
-import { defaultWellAttributes } from '@/stores/utilities/run'
-import { usePacbioRunCreateStore } from '@/stores/pacbioRunCreate'
+import { defaultWellAttributes } from '@/stores/utilities/run.js'
+import { usePacbioRunCreateStore } from '@/stores/pacbioRunCreateV1.js'
 
 // required as suggestion to remove the deprecated function
 // https://vue-test-utils.vuejs.org/api/options.html#attachtodocument
