+++ tests/unit/stores/pacbioRunCreateV1.spec.js
@@ -1,8 +1,8 @@
 import { createPinia, setActivePinia } from 'pinia'
-import { usePacbioRunCreateStore } from '@/stores/pacbioRunCreate'
+import { usePacbioRunCreateStore } from '@/stores/pacbioRunCreateV1.js'
 import useRootStore from '@/stores'
-import storePools from '@tests/data/StoreRunPools'
-import { Data } from '@support/testHelper'
+import storePools from '@tests/data/StoreRunPools.json'
+import { Data } from '@support/testHelper.js'
 import * as jsonapi from '@/api/JsonApi'
 import {
   newRun,
